+++ ChangeLog.md
@@ -1,6 +1,10 @@
 ﻿# ChangeLog
 KsWare.Presentation.Lite
 
+#### [2020-01-27]
+- ViewLocator, IViewLocatorStrategy, ...
+- ViewModelPresenter, ViewModelViewConverter
+
 #### v0.1 [2020-01-26]
 - NotifyPropertyChangedBase
 - BootstrapperBase